Book_details default format
The book_details should be in this formatting. A Type error or Value Error might occur after running the program after closing it.
The issue still remains of over writing the file contents by wrapping them as a string in a list. With more iteration quotations "" and brackets [[]] are added to the text file, which makes the borrow process and returns process errors some